requirements.txt
torch==1.3.0
advertorch==0.2.0
tensorboardX==1.9
cupy==6.5.0
torchvision==0.2.1
tqdm==4.36.1
Pillow==6.2.0
numpy==1.17.2